+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## 1.5.0 - 2016-05-05
 ### Added
 - Shortcut `-i` to invoke `--ignore-delays` option of `run` command. ([#69](https://github.com/lmc-eu/steward/pull/69))
+- Command `results` to show test results summary from the command line (CLI equivalent to viewing `results.xml` in a browser).
 
 ### Changed
 - Prepare removal of deprecated features - show deprecation warning when `run-tests` alias is being used. Note this alias and also non-namespaced WebDriver (deprecated in 1.2.0) will be removed in future Steward versions! ([#60](https://github.com/lmc-eu/steward/pull/60))